requirements.txt
editdistance
pillow
rouge
datasets
torchvision==0.16.2
deepspeed
icetk
ftfy
hjson
ninja
tokenizers
transformers>=4.38.2
accelerate>=0.23.0
datasets==2.16.1
wcwidth==0.2.12
pyarrow==12.0.1
jieba
sentencepiece
rjieba
sqlitedict
sacrebleu
datasets
tensorboard
lm_dataformat
tiktoken
